Split up thing.ts in thing/get and thing/remove
thing.ts was getting quite unwieldy, and although all functions
operate on Things, they are relatively loosely coupled. This puts
code that is likely to be changed together, together.
