+++ src/thing.ts
@@ -1,4 +1,4 @@
-import { NamedNode, Quad, Literal, Quad_Object } from "rdf-js";
+import { NamedNode } from "rdf-js";
 import {
   LitDataset,
   IriString,
@@ -12,12 +12,11 @@ import {
   hasDiff,
   hasMetadata,
 } from "./index";
-import { dataset, DataFactory, filter } from "./rdfjs";
+import { dataset } from "./rdfjs";
 import {
   isLocalNode,
   isEqual,
   isNamedNode,
-  isLiteral,
   getLocalNode,
   asNamedNode,
   resolveLocalIri,
@@ -235,830 +234,6 @@ export function asIri(thing: Thing, baseIri?: IriString): IriString {
   }
 
   return thing.iri;
-}
-
-function filterThing<T extends Thing>(
-  thing: T,
-  callback: (quad: Quad) => boolean
-): T extends ThingLocal ? ThingLocal : ThingPersisted;
-function filterThing(thing: Thing, callback: (quad: Quad) => boolean): Thing {
-  const filtered = filter(thing, callback);
-  if (isThingLocal(thing)) {
-    (filtered as ThingLocal).name = thing.name;
-    return filtered as ThingLocal;
-  }
-  (filtered as ThingPersisted).iri = thing.iri;
-  return filtered as ThingPersisted;
-}
-
-/**
- * @param thing The [[Thing]] to read an IRI value from.
- * @param predicate The given Predicate for which you want the IRI value.
- * @returns An IRI value for the given Predicate, if present, or null otherwise.
- */
-export function getOneIri(
-  thing: Thing,
-  predicate: Iri | IriString
-): IriString | null {
-  const namedNodeMatcher = getNamedNodeMatcher(predicate);
-
-  const matchingQuad = findOne(thing, namedNodeMatcher);
-
-  if (matchingQuad === null) {
-    return null;
-  }
-
-  return matchingQuad.object.value;
-}
-
-/**
- * @param thing The [[Thing]] to read the IRI values from.
- * @param predicate The given Predicate for which you want the IRI values.
- * @returns The IRI values for the given Predicate.
- */
-export function getAllIris(
-  thing: Thing,
-  predicate: Iri | IriString
-): IriString[] {
-  const iriMatcher = getNamedNodeMatcher(predicate);
-
-  const matchingQuads = findAll(thing, iriMatcher);
-
-  return matchingQuads.map((quad) => quad.object.value);
-}
-
-/**
- * @param thing Thing to remove an IRI value from.
- * @param predicate Predicate for which to remove the given IRI value.
- * @param value IRI to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export function removeOneIri<T extends Thing>(
-  thing: T,
-  predicate: Iri | IriString,
-  value: Iri | IriString | ThingPersisted
-): T extends ThingLocal ? ThingLocal : ThingPersisted;
-export function removeOneIri(
-  thing: Thing,
-  predicate: Iri | IriString,
-  value: Iri | IriString | ThingPersisted
-): Thing {
-  const predicateNode = asNamedNode(predicate);
-  const iriNode = isNamedNode(value)
-    ? value
-    : typeof value === "string"
-    ? asNamedNode(value)
-    : asNamedNode(asIri(value));
-
-  const updatedThing = filterThing(thing, (quad) => {
-    return (
-      !quad.predicate.equals(predicateNode) ||
-      !isNamedNode(quad.object) ||
-      !quad.object.equals(iriNode)
-    );
-  });
-  return updatedThing;
-}
-
-function removeOneLiteralOfType<T extends Thing>(
-  thing: T,
-  predicate: Iri | IriString,
-  value: string,
-  type: IriString
-): T extends ThingLocal ? ThingLocal : ThingPersisted;
-function removeOneLiteralOfType(
-  thing: Thing,
-  predicate: Iri | IriString,
-  value: string,
-  type: IriString
-): Thing {
-  const updatedThing = removeOneLiteral(
-    thing,
-    predicate,
-    DataFactory.literal(value, DataFactory.namedNode(type))
-  );
-  return updatedThing;
-}
-
-/**
- * @param thing Thing to remove a value from.
- * @param predicate Predicate for which to remove the given value.
- * @param value Value to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-type RemoveOneOfType<Type> = <T extends Thing>(
-  thing: T,
-  predicate: Iri | IriString,
-  value: Type
-) => T extends ThingLocal ? ThingLocal : ThingPersisted;
-
-/**
- * @param thing The [[Thing]] to read a string value from.
- * @param predicate The given Predicate for which you want the string value.
- * @returns A string value for the given Predicate, if present, or null otherwise.
- */
-export function getOneStringUnlocalised(
-  thing: Thing,
-  predicate: Iri | IriString
-): string | null {
-  const literalString = getOneLiteralOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#string"
-  );
-
-  return literalString;
-}
-
-/**
- * @param thing The [[Thing]] to read the string values from.
- * @param predicate The given Predicate for which you want the string values.
- * @returns The string values for the given Predicate.
- */
-export function getAllStringUnlocaliseds(
-  thing: Thing,
-  predicate: Iri | IriString
-): string[] {
-  const literalStrings = getAllLiteralsOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#string"
-  );
-
-  return literalStrings;
-}
-
-/**
- * @param thing Thing to remove an unlocalised string value from.
- * @param predicate Predicate for which to remove the given string value.
- * @param value String to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export const removeOneStringUnlocalised: RemoveOneOfType<string> = (
-  thing,
-  predicate,
-  value
-) => {
-  return removeOneLiteralOfType(
-    thing,
-    predicate,
-    value,
-    "http://www.w3.org/2001/XMLSchema#string"
-  );
-};
-
-type LiteralLocaleString = Literal & {
-  datatype: { value: "http://www.w3.org/1999/02/22-rdf-syntax-ns#langString" };
-  language: string;
-};
-const getLocaleStringMatcher = function (
-  predicate: Iri | IriString,
-  locale: string
-): Matcher<LiteralLocaleString> {
-  const predicateNode = asNamedNode(predicate);
-
-  const matcher = function (
-    quad: Quad
-  ): quad is QuadWithObject<LiteralLocaleString> {
-    return (
-      predicateNode.equals(quad.predicate) &&
-      isLiteral(quad.object) &&
-      quad.object.datatype.value ===
-        "http://www.w3.org/1999/02/22-rdf-syntax-ns#langString" &&
-      quad.object.language.toLowerCase() === locale.toLowerCase()
-    );
-  };
-  return matcher;
-};
-
-/**
- * @param thing The [[Thing]] to read a localised string value from.
- * @param predicate The given Predicate for which you want the localised string value.
- * @param locale The desired locale for the string value.
- * @returns A localised string value for the given Predicate, if present in `locale`, or null otherwise.
- */
-export function getOneStringInLocale(
-  thing: Thing,
-  predicate: Iri | IriString,
-  locale: string
-): string | null {
-  const localeStringMatcher = getLocaleStringMatcher(predicate, locale);
-
-  const matchingQuad = findOne(thing, localeStringMatcher);
-
-  if (matchingQuad === null) {
-    return null;
-  }
-
-  return matchingQuad.object.value;
-}
-
-/**
- * @param thing The [[Thing]] to read the localised string values from.
- * @param predicate The given Predicate for which you want the localised string values.
- * @param locale The desired locale for the string values.
- * @returns The localised string values for the given Predicate.
- */
-export function getAllStringsInLocale(
-  thing: Thing,
-  predicate: Iri | IriString,
-  locale: string
-): string[] {
-  const localeStringMatcher = getLocaleStringMatcher(predicate, locale);
-
-  const matchingQuads = findAll(thing, localeStringMatcher);
-
-  return matchingQuads.map((quad) => quad.object.value);
-}
-
-/**
- * @param thing Thing to remove a localised string value from.
- * @param predicate Predicate for which to remove the given localised string value.
- * @param value String to remove from `thing` for the given `predicate`.
- * @param locale Locale of the string to remove.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export function removeOneStringInLocale<T extends Thing>(
-  thing: T,
-  predicate: Iri | IriString,
-  value: string,
-  locale: string
-): T extends ThingLocal ? ThingLocal : ThingPersisted;
-export function removeOneStringInLocale(
-  thing: Thing,
-  predicate: Iri | IriString,
-  value: string,
-  locale: string
-): Thing {
-  // Note: Due to how the `DataFactory.literal` constructor behaves, this function
-  // must call directly `removeOneLiteral` directly, with the locale as the data
-  // type of the Literal (which is not a valid NamedNode).
-  return removeOneLiteral(
-    thing,
-    predicate,
-    DataFactory.literal(value, locale.toLowerCase())
-  );
-}
-
-/**
- * @param thing The [[Thing]] to read an integer value from.
- * @param predicate The given Predicate for which you want the integer value.
- * @returns An integer value for the given Predicate, if present, or null otherwise.
- */
-export function getOneInteger(
-  thing: Thing,
-  predicate: Iri | IriString
-): number | null {
-  const literalString = getOneLiteralOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#integer"
-  );
-
-  if (literalString === null) {
-    return null;
-  }
-
-  return Number.parseInt(literalString, 10);
-}
-
-/**
- * @param thing The [[Thing]] to read the integer values from.
- * @param predicate The given Predicate for which you want the integer values.
- * @returns The integer values for the given Predicate.
- */
-export function getAllIntegers(
-  thing: Thing,
-  predicate: Iri | IriString
-): number[] {
-  const literalStrings = getAllLiteralsOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#integer"
-  );
-
-  return literalStrings.map((literalString) =>
-    Number.parseInt(literalString, 10)
-  );
-}
-
-/**
- * @param thing Thing to remove an integer value from.
- * @param predicate Predicate for which to remove the given integer value.
- * @param value Integer to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export const removeOneInteger: RemoveOneOfType<number> = (
-  thing,
-  predicate,
-  value
-) => {
-  return removeOneLiteralOfType(
-    thing,
-    predicate,
-    value.toString(),
-    "http://www.w3.org/2001/XMLSchema#integer"
-  );
-};
-
-/**
- * @param thing The [[Thing]] to read a decimal value from.
- * @param predicate The given Predicate for which you want the decimal value.
- * @returns A decimal value for the given Predicate, if present, or null otherwise.
- */
-export function getOneDecimal(
-  thing: Thing,
-  predicate: Iri | IriString
-): number | null {
-  const literalString = getOneLiteralOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#decimal"
-  );
-
-  if (literalString === null) {
-    return null;
-  }
-
-  return Number.parseFloat(literalString);
-}
-
-/**
- * @param thing The [[Thing]] to read the decimal values from.
- * @param predicate The given Predicate for which you want the decimal values.
- * @returns The decimal values for the given Predicate.
- */
-export function getAllDecimals(
-  thing: Thing,
-  predicate: Iri | IriString
-): number[] {
-  const literalStrings = getAllLiteralsOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#decimal"
-  );
-
-  return literalStrings.map((literalString) =>
-    Number.parseFloat(literalString)
-  );
-}
-
-/**
- * @param thing Thing to remove a decimal value from.
- * @param predicate Predicate for which to remove the given decimal value.
- * @param value Decimal to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export const removeOneDecimal: RemoveOneOfType<number> = (
-  thing,
-  predicate,
-  value
-) => {
-  return removeOneLiteralOfType(
-    thing,
-    predicate,
-    value.toString(),
-    "http://www.w3.org/2001/XMLSchema#decimal"
-  );
-};
-
-/**
- * @param thing The [[Thing]] to read a boolean value from.
- * @param predicate The given Predicate for which you want the boolean value.
- * @returns A boolean value for the given Predicate, if present, or null otherwise.
- */
-export function getOneBoolean(
-  thing: Thing,
-  predicate: Iri | IriString
-): boolean | null {
-  const literalString = getOneLiteralOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#boolean"
-  );
-
-  if (literalString === null) {
-    return null;
-  }
-
-  if (literalString === "1") {
-    return true;
-  } else if (literalString === "0") {
-    return false;
-  } else {
-    return null;
-  }
-}
-
-/**
- * @param thing The [[Thing]] to read the boolean values from.
- * @param predicate The given Predicate for which you want the boolean values.
- * @returns The boolean values for the given Predicate.
- */
-export function getAllBooleans(
-  thing: Thing,
-  predicate: Iri | IriString
-): boolean[] {
-  const literalStrings = getAllLiteralsOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#boolean"
-  );
-
-  return literalStrings
-    .map((literalString) => {
-      if (literalString === "1") {
-        return true;
-      } else if (literalString === "0") {
-        return false;
-      } else {
-        return null;
-      }
-    })
-    .filter((possibleBoolean) => possibleBoolean !== null) as boolean[];
-}
-
-/**
- * @param thing Thing to remove a boolean value from.
- * @param predicate Predicate for which to remove the given boolean value.
- * @param value Boolean to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export const removeOneBoolean: RemoveOneOfType<boolean> = (
-  thing,
-  predicate,
-  value
-) => {
-  return removeOneLiteralOfType(
-    thing,
-    predicate,
-    value ? "1" : "0",
-    "http://www.w3.org/2001/XMLSchema#boolean"
-  );
-};
-
-/**
- * @param thing The [[Thing]] to read a datetime value from.
- * @param predicate The given Predicate for which you want the datetime value.
- * @returns A datetime value for the given Predicate, if present, or null otherwise.
- */
-export function getOneDatetime(
-  thing: Thing,
-  predicate: Iri | IriString
-): Date | null {
-  const literalString = getOneLiteralOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#dateTime"
-  );
-
-  if (literalString === null) {
-    return null;
-  }
-
-  return parseDatetimeFromLiteralString(literalString);
-}
-
-/**
- * @param thing The [[Thing]] to read the datetime values from.
- * @param predicate The given Predicate for which you want the datetime values.
- * @returns The datetime values for the given Predicate.
- */
-export function getAllDatetimes(
-  thing: Thing,
-  predicate: Iri | IriString
-): Date[] {
-  const literalStrings = getAllLiteralsOfType(
-    thing,
-    predicate,
-    "http://www.w3.org/2001/XMLSchema#dateTime"
-  );
-
-  return literalStrings
-    .map(parseDatetimeFromLiteralString)
-    .filter((potentialDatetime) => potentialDatetime !== null) as Date[];
-}
-
-function parseDatetimeFromLiteralString(literalString: string): Date | null {
-  if (
-    literalString === null ||
-    literalString.length <= 17 ||
-    literalString.indexOf("Z") === -1
-  ) {
-    return null;
-  }
-
-  // See https://github.com/linkeddata/rdflib.js/blob/d84af88f367b8b5f617c753d8241c5a2035458e8/src/literal.js#L87
-  const utcFullYear = parseInt(literalString.substring(0, 4), 10);
-  const utcMonth = parseInt(literalString.substring(5, 7), 10) - 1;
-  const utcDate = parseInt(literalString.substring(8, 10), 10);
-  const utcHours = parseInt(literalString.substring(11, 13), 10);
-  const utcMinutes = parseInt(literalString.substring(14, 16), 10);
-  const utcSeconds = parseInt(
-    literalString.substring(17, literalString.indexOf("Z")),
-    10
-  );
-  const date = new Date(0);
-  date.setUTCFullYear(utcFullYear);
-  date.setUTCMonth(utcMonth);
-  date.setUTCDate(utcDate);
-  date.setUTCHours(utcHours);
-  date.setUTCMinutes(utcMinutes);
-  date.setUTCSeconds(utcSeconds);
-  return date;
-}
-
-/**
- * @param thing Thing to remove a datetime value from.
- * @param predicate Predicate for which to remove the given datetime value.
- * @param value Datetime to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export const removeOneDatetime: RemoveOneOfType<Date> = (
-  thing,
-  predicate,
-  value
-) => {
-  return removeOneLiteralOfType(
-    thing,
-    predicate,
-    serialiseDatetime(value),
-    "http://www.w3.org/2001/XMLSchema#dateTime"
-  );
-};
-
-function serialiseDatetime(datetime: Date): string {
-  // To align with rdflib, we ignore miliseconds:
-  // https://github.com/linkeddata/rdflib.js/blob/d84af88f367b8b5f617c753d8241c5a2035458e8/src/literal.js#L74
-  const roundedDate = new Date(
-    Date.UTC(
-      datetime.getUTCFullYear(),
-      datetime.getUTCMonth(),
-      datetime.getUTCDate(),
-      datetime.getUTCHours(),
-      datetime.getUTCMinutes(),
-      datetime.getUTCSeconds(),
-      0
-    )
-  );
-  // Truncate the `.000Z` at the end (i.e. the miliseconds), to plain `Z`:
-  const rdflibStyleString = roundedDate.toISOString().replace(/\.000Z$/, "Z");
-  return rdflibStyleString;
-}
-
-const getLiteralMatcher = function (
-  predicate: Iri | IriString
-): Matcher<Literal> {
-  const predicateNode = asNamedNode(predicate);
-
-  const matcher = function (quad: Quad): quad is QuadWithObject<Literal> {
-    return predicateNode.equals(quad.predicate) && isLiteral(quad.object);
-  };
-  return matcher;
-};
-
-/**
- * @param thing The [[Thing]] to read a Literal value from.
- * @param predicate The given Predicate for which you want the Literal value.
- * @returns A Literal value for the given Predicate, if present, or null otherwise.
- * @ignore This should not be needed due to the other getOne*() functions. If you do find yourself needing it, please file a feature request for your use case.
- */
-export function getOneLiteral(
-  thing: Thing,
-  predicate: Iri | IriString
-): Literal | null {
-  const literalMatcher = getLiteralMatcher(predicate);
-
-  const matchingQuad = findOne(thing, literalMatcher);
-
-  if (matchingQuad === null) {
-    return null;
-  }
-
-  return matchingQuad.object;
-}
-
-/**
- * @param thing The [[Thing]] to read the Literal values from.
- * @param predicate The given Predicate for which you want the Literal values.
- * @returns The Literal values for the given Predicate.
- * @ignore This should not be needed due to the other getAll*() functions. If you do find yourself needing it, please file a feature request for your use case.
- */
-export function getAllLiterals(
-  thing: Thing,
-  predicate: Iri | IriString
-): Literal[] {
-  const literalMatcher = getLiteralMatcher(predicate);
-
-  const matchingQuads = findAll(thing, literalMatcher);
-
-  return matchingQuads.map((quad) => quad.object);
-}
-
-/**
- * @ignore This should not be needed due to the other removeOne*() functions. If you do find yourself needing it, please file a feature request for your use case.
- * @param thing Thing to remove a Literal value from.
- * @param predicate Predicate for which to remove the given Literal value.
- * @param value Literal to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export function removeOneLiteral<T extends Thing>(
-  thing: T,
-  predicate: Iri | IriString,
-  value: Literal
-): T extends ThingLocal ? ThingLocal : ThingPersisted;
-export function removeOneLiteral(
-  thing: Thing,
-  predicate: Iri | IriString,
-  value: Literal
-): Thing {
-  const predicateNode = asNamedNode(predicate);
-  const updatedThing = filterThing(thing, (quad) => {
-    return (
-      !quad.predicate.equals(predicateNode) ||
-      !isLiteral(quad.object) ||
-      !quad.object.equals(value)
-    );
-  });
-  return updatedThing;
-}
-
-const getNamedNodeMatcher = function (
-  predicate: Iri | IriString
-): Matcher<NamedNode> {
-  const predicateNode = asNamedNode(predicate);
-
-  const matcher = function (quad: Quad): quad is QuadWithObject<NamedNode> {
-    return predicateNode.equals(quad.predicate) && isNamedNode(quad.object);
-  };
-  return matcher;
-};
-
-/**
- * @param thing The [[Thing]] to read a NamedNode value from.
- * @param predicate The given Predicate for which you want the NamedNode value.
- * @returns A NamedNode value for the given Predicate, if present, or null otherwise.
- * @ignore This should not be needed due to the other getOne*() functions. If you do find yourself needing it, please file a feature request for your use case.
- */
-export function getOneNamedNode(
-  thing: Thing,
-  predicate: Iri | IriString
-): NamedNode | null {
-  const namedNodeMatcher = getNamedNodeMatcher(predicate);
-
-  const matchingQuad = findOne(thing, namedNodeMatcher);
-
-  if (matchingQuad === null) {
-    return null;
-  }
-
-  return matchingQuad.object;
-}
-
-/**
- * @param thing The [[Thing]] to read the NamedNode values from.
- * @param predicate The given Predicate for which you want the NamedNode values.
- * @returns The NamedNode values for the given Predicate.
- * @ignore This should not be needed due to the other getOne*() functions. If you do find yourself needing it, please file a feature request for your use case.
- */
-export function getAllNamedNodes(
-  thing: Thing,
-  predicate: Iri | IriString
-): NamedNode[] {
-  const namedNodeMatcher = getNamedNodeMatcher(predicate);
-
-  const matchingQuads = findAll(thing, namedNodeMatcher);
-
-  return matchingQuads.map((quad) => quad.object);
-}
-
-/**
- * @ignore This should not be needed due to the other removeOne*() functions. If you do find yourself needing it, please file a feature request for your use case.
- * @param thing Thing to remove a NamedNode value from.
- * @param predicate Predicate for which to remove the given NamedNode value.
- * @param value NamedNode to remove from `thing` for the given `predicate`.
- * @returns A new Thing equal to the input Thing with the given value removed for the given Predicate.
- */
-export function removeOneNamedNode<T extends Thing>(
-  thing: T,
-  predicate: Iri | IriString,
-  value: NamedNode
-): T extends ThingLocal ? ThingLocal : ThingPersisted;
-export function removeOneNamedNode(
-  thing: Thing,
-  predicate: Iri | IriString,
-  value: NamedNode
-): Thing {
-  const predicateNode = asNamedNode(predicate);
-  const updatedThing = filterThing(thing, (quad) => {
-    return (
-      !quad.predicate.equals(predicateNode) ||
-      !isNamedNode(quad.object) ||
-      !quad.object.equals(value)
-    );
-  });
-  return updatedThing;
-}
-
-type LiteralOfType<Type extends IriString> = Literal & {
-  datatype: { value: Type };
-};
-const getLiteralOfTypeMatcher = function <Datatype extends IriString>(
-  predicate: Iri | IriString,
-  datatype: Datatype
-): Matcher<LiteralOfType<Datatype>> {
-  const predicateNode = asNamedNode(predicate);
-
-  const matcher = function (
-    quad: Quad
-  ): quad is QuadWithObject<LiteralOfType<Datatype>> {
-    return (
-      predicateNode.equals(quad.predicate) &&
-      isLiteral(quad.object) &&
-      quad.object.datatype.value === datatype
-    );
-  };
-  return matcher;
-};
-
-/**
- * @param thing The [Thing]] to read a Literal of the given type from.
- * @param predicate The given Predicate for which you want the Literal value.
- * @param literalType Set type of the Literal data.
- * @returns The stringified value for the given Predicate and type, if present, or null otherwise.
- */
-function getOneLiteralOfType<Datatype extends IriString>(
-  thing: Thing,
-  predicate: Iri | IriString,
-  literalType: Datatype
-): string | null {
-  const literalOfTypeMatcher = getLiteralOfTypeMatcher(predicate, literalType);
-
-  const matchingQuad = findOne(thing, literalOfTypeMatcher);
-
-  if (matchingQuad === null) {
-    return null;
-  }
-
-  return matchingQuad.object.value;
-}
-
-/**
- * @param thing The [Thing]] to read the Literals of the given type from.
- * @param predicate The given Predicate for which you want the Literal values.
- * @param literalType Set type of the Literal data.
- * @returns The stringified values for the given Predicate and type.
- */
-function getAllLiteralsOfType<Datatype extends IriString>(
-  thing: Thing,
-  predicate: Iri | IriString,
-  literalType: Datatype
-): string[] {
-  const literalOfTypeMatcher = getLiteralOfTypeMatcher(predicate, literalType);
-
-  const matchingQuads = findAll(thing, literalOfTypeMatcher);
-
-  return matchingQuads.map((quad) => quad.object.value);
-}
-
-type QuadWithObject<Object extends Quad_Object> = Quad & { object: Object };
-type Matcher<Object extends Quad_Object> = (
-  quad: Quad
-) => quad is QuadWithObject<Object>;
-
-/**
- * @param thing The [[Thing]] to extract a Quad from.
- * @param matcher Callback function that returns a boolean indicating whether a given Quad should be included.
- * @returns First Quad in `thing` for which `matcher` returned true.
- */
-function findOne<Object extends Quad_Object>(
-  thing: Thing,
-  matcher: Matcher<Object>
-): QuadWithObject<Object> | null {
-  for (let quad of thing) {
-    if (matcher(quad)) {
-      return quad;
-    }
-  }
-  return null;
-}
-
-/**
- * @param thing The [[Thing]] to extract Quads from.
- * @param matcher Callback function that returns a boolean indicating whether a given Quad should be included.
- * @returns All Quads in `thing` for which `matcher` returned true.
- */
-function findAll<Object extends Quad_Object>(
-  thing: Thing,
-  matcher: Matcher<Object>
-): QuadWithObject<Object>[] {
-  const matched: QuadWithObject<Object>[] = [];
-  for (let quad of thing) {
-    if (matcher(quad)) {
-      matched.push(quad);
-    }
-  }
-  return matched;
 }
 
 /**
